Update module name to raincatcher-sync
Motivation:
Currently this project's module name is named after the old Work Force
Management (WFM) which has changed to Raincatcher. But the npm module
names have not been updated which is that this commit does.

Modifications:
Updated the module name in package.json and updated README.md to reflect
the module change.

JIRA:
https://issues.jboss.org/browse/RAINCATCH-230